Turn `math.floor()` resut into int
This seems as use case for `random.randrange()`.

+++ tests/test_inlet_outlet.py
@@ -30,8 +30,8 @@ def main(show_plot=True, time_steps=600, plot_freq=60):
 
     planes = ('XY', 'XZ', 'YZ')
     tests  = array([11,12,13,14, 21,22,23,24, 31,32,33,34, 41,42,43,44])
-    TEST   = tests[ floor(random()*16) ]
-    PLANE  = planes[ floor(random()*3) ]
+    TEST   = tests[int(floor(random() * 16))]
+    PLANE  = planes[int(floor(random() * 3))]
 
     TEST = 23
